refactor: add auth middleware to file route
