+++ src/files/models.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 from datetime import datetime
 
 from sqlalchemy import String, Text, DateTime, func
-from sqlalchemy.orm import declarative_base, Mapped, mapped_column
+from sqlalchemy.orm import Mapped, mapped_column
 
 from cuid2 import cuid_wrapper
 
-Base = declarative_base()
+from src.database import Base
 
 
 class FileObject(Base):
